+++ package.json
@@ -82,7 +82,6 @@
     "cheminfo-font": "^1.9.0",
     "cross-env": "^7.0.3",
     "eslint": "^8.25.0",
-    "file-saver": "^2.0.5",
     "ml-signal-processing": "^0.5.2",
     "ml-spectra-processing": "^11.13.0",
     "prettier": "^2.7.1",
